feat: add activity service layer, switch points/unit fields to integers

+++ api/app/schemas/category.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel
 class CategoryResponse(BaseModel):
     id: int
     name: str
-    points_per_unit: Decimal
+    points_per_unit: int
     color: str | None
     icon: str | None
